Split example into linking, compression examples.
