+++ examples/README.md
@@ -1,0 +1,9 @@
+# Examples
+
+This directory contains examples showing how to use MGARD.
+
+1. [an example showing to how link against MGARD][linking]
+2. [an example showing how to compress with MGARD][compression]
+
+[linking]: examples/linking/README.md
+[compression]: examples/compression/README.md
