--- examples/basic/README.md
@@ -1,11 +1,0 @@
-# Basic Compression Example
-
-To build and run this example, build and install MGARD and then run the following in the directory containing this file.
-
-```console
-$ cmake -S . -B build
-$ cmake --build build
-$ build/main
-```
-
-`CMakeLists.txt` contains some hints if you encounter problems finding dependencies.
