+++ examples/linking/src/main.cpp
@@ -1,0 +1,8 @@
+#include <iostream>
+
+#include "mgard/compress.hpp"
+
+int main() {
+  std::cout << "you have successfully linked against MGARD" << std::endl;
+  return 0;
+}
